阶段性1

+++ Module/data_type.h
@@ -68,7 +68,7 @@ typedef struct {
 
 typedef struct {
   uint8_t data[4];
-  uint32_t   distance;
+  float   distance;
 } pub_Encoder_Data;
 
 #pragma pack()
